Re-orient servlet asyncrony to trigger when enter returns nil.
This is an intermediate step towards fully removing pause and
with-pause.

Previously, asynchronous servlet behavior was triggered by the
pause-stylobate function being called from with-pause. Now, instead,
we set a servlet as asynchronous when a nil return value is
encountered on an enter function. This is accomplished by adding a
start-servlet-async function to the :async-enter function (in
enter-stylobate). When a nil context is encountered all functions in
the :async-enter vector are executed and our servlet is set to
asynchronous.